+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/business-licence-application/components/business-controlling-members-and-employees.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { BusinessLicenceApplicationRoutes } from '../business-licence-application-routing.module';
+import { BusinessLicenceApplicationRoutes } from '@app/modules/business-licence-application/business-license-application-routes';
+
 
 @Component({
 	selector: 'app-business-controlling-members-and-employees',
