+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/business-licence-application/components/business-licence-application-base.component.ts
@@ -4,7 +4,8 @@ import { AuthProcessService } from '@app/core/services/auth-process.service';
 import { AuthUserBceidService } from '@app/core/services/auth-user-bceid.service';
 import { BusinessApplicationService } from '@app/core/services/business-application.service';
 import { lastValueFrom, take, tap } from 'rxjs';
-import { BusinessLicenceApplicationRoutes } from '../business-licence-application-routing.module';
+import { BusinessLicenceApplicationRoutes } from '@app/modules/business-licence-application/business-license-application-routes';
+
 
 @Component({
 	selector: 'app-business-licence-application-base',
