+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/business-licence-application/components/business-managers.component.ts
@@ -7,7 +7,8 @@ import { AuthUserBceidService } from '@app/core/services/auth-user-bceid.service
 import { BusinessApplicationService } from '@app/core/services/business-application.service';
 import { DialogComponent, DialogOptions } from '@app/shared/components/dialog.component';
 import { HotToastService } from '@ngxpert/hot-toast';
-import { BusinessLicenceApplicationRoutes } from '../business-licence-application-routing.module';
+import { BusinessLicenceApplicationRoutes } from '@app/modules/business-licence-application/business-license-application-routes';
+
 import { BizPortalUserDialogData, ModalBusinessManagerEditComponent } from './modal-business-manager-edit.component';
 
 @Component({
