+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/business-licence-application/components/business-profile.component.ts
@@ -3,7 +3,8 @@ import { Router } from '@angular/router';
 import { BusinessApplicationService } from '@app/core/services/business-application.service';
 import { UtilService } from '@app/core/services/util.service';
 import { HotToastService } from '@ngxpert/hot-toast';
-import { BusinessLicenceApplicationRoutes } from '../business-licence-application-routing.module';
+import { BusinessLicenceApplicationRoutes } from '@app/modules/business-licence-application/business-license-application-routes';
+
 import { CommonBusinessProfileComponent } from './common-business-profile.component';
 
 @Component({
